Refactored session-core functions for container creation to new file called pod-helper; Moved pod-helper into folder called pod-management under utils; Moved unit tests for these functions to pod-helper.test.js; Created new unit test for setDocContainerAclPermission in session-core.test.js
